Add new non-blocking assemble_scans2_publish topic
Add a new non-blocking path for assembling scans directly to a topic.
The topic name is "cloud". The request comes in on a topic, not a
service, so the caller is not blocked in any way while the request is
processed. The requests are the same style as the AssembleScans2Blocking
service, so the request is not immediately published if the blocking
period is set (to allow the request to wait for all scans to be present
prior to assembly).